refactor: remove `hydroflow::lang` module, move `Clear`, `MonotonicMap` to `hydroflow::util` instead (#677)
Refs: #675
