+++ hydroflow_lang/src/graph/ops/enumerate.rs
@@ -75,7 +75,7 @@ pub const ENUMERATE: OperatorConstraints = OperatorConstraints {
             Persistence::Tick => (
                 quote_spanned! {op_span=>
                     let #counter_ident = #hydroflow.add_state(::std::cell::RefCell::new(
-                        #root::lang::monotonic_map::MonotonicMap::new_init(0..),
+                        #root::util::monotonic_map::MonotonicMap::new_init(0..),
                     ));
                 },
                 quote_spanned! {op_span=>
