+++ hydroflow_lang/src/graph/ops/lattice_join.rs
@@ -127,7 +127,7 @@ pub const LATTICE_JOIN: OperatorConstraints = OperatorConstraints {
             let (init, borrow) = match persistence {
                 Persistence::Tick => (
                     quote_spanned! {op_span=>
-                        #root::lang::monotonic_map::MonotonicMap::new_init(
+                        #root::util::monotonic_map::MonotonicMap::new_init(
                             #root::compiled::pull::HalfJoinStateLattice::default()
                         )
                     },
